Pulled out methodModel into own file.  Added utils for const
